CSC4350/project3/requirements.txt
Flask
requests
python-dotenv
flask-login
Flask-SQLAlchemy
psycopg2-binary
